+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Syntax: Allow `if` without `else`, as well as `else if (...) { ... }`.
 
 Interactive interpreter: added `:abort` and `:quit` commands.
 
-# 0.2 (released 14 March 2023)
+# 0.2: FizzBuzz (released 14 March 2023)
 
 Syntax: Added `if`, `while`, assignment, `!=`, `-`, `*`, `/`, `<`,
 `>`, `||` and `&&`.
@@ -14,6 +14,6 @@ Standard library: Added `int_to_string`.
 Interactive interpreter: added `:stack` and `:source`
 commands. Commands can now be run when errors occur.
 
-# 0.1 (released 11 February 2023)
+# 0.1: Hello World (released 11 February 2023)
 
 A basic toplevel that supports hello world.